+++ Time-Up/App.tsx
@@ -5,26 +5,26 @@ import { Dimensions, Platform, useWindowDimensions } from 'react-native';
 import { Provider as PaperProvider } from 'react-native-paper';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import './global.css';
-import AlarmMemoPage from './src/pages/AlarmMemoPage';
-import AlarmPage from './src/pages/AlarmPage';
+import AlarmMemoPage from './src/pages/Alarm/AlarmMemoPage';
+import AlarmPage from './src/pages/Alarm/AlarmPage';
+import EditMyAlarmPage from './src/pages/Alarm/EditMyAlarmPage';
+import EditWakeUpAlarmPage from './src/pages/Alarm/EditWakeUpAlarmPage';
+import MyAlarmDetailPage from './src/pages/Alarm/MyAlarmDetailPage';
+import MyAlarmPage from './src/pages/Alarm/MyAlarmPage';
+import PushAlarmPage from './src/pages/Alarm/PushAlarmPage';
+import SelectAlarmDatePage from './src/pages/Alarm/SelectAlarmDatePage';
+import SelectAlarmReplayPage from './src/pages/Alarm/SelectAlarmReplayPage';
+import SelectAlarmSoundPage from './src/pages/Alarm/SelectAlarmSoundPage';
+import SelectAlarmTimePage from './src/pages/Alarm/SelectAlarmTimePage';
+import SelectAlarmVibratePage from './src/pages/Alarm/SelectAlarmVibratePage';
+import WakeUpAlarmDetail from './src/pages/Alarm/WakeUpAlarmDetail';
+import WakeUpAlarmPage from './src/pages/Alarm/WakeUpAlarmPage';
+import WakeUpAlarmPageNotAuto from './src/pages/Alarm/WakeUpAlarmPageNotAuto';
 import CalendarPage from './src/pages/CalendarPage';
 import DiaryWritePage from './src/pages/DiaryWritePage';
-import EditMyAlarmPage from './src/pages/EditMyAlarmPage';
-import EditWakeUpAlarmPage from './src/pages/EditWakeUpAlarmPage';
 import LoginPage from './src/pages/LoginPage';
-import MyAlarmDetailPage from './src/pages/MyAlarmDetailPage';
-import MyAlarmPage from './src/pages/MyAlarmPage';
 import MyPage from './src/pages/MyPage';
 import OnboardingPage from './src/pages/OnboardingPage';
-import PushAlarmPage from './src/pages/PushAlarmPage';
-import SelectAlarmDatePage from './src/pages/SelectAlarmDatePage';
-import SelectAlarmReplayPage from './src/pages/SelectAlarmReplayPage';
-import SelectAlarmSoundPage from './src/pages/SelectAlarmSoundPage';
-import SelectAlarmTimePage from './src/pages/SelectAlarmTimePage';
-import SelectAlarmVibratePage from './src/pages/SelectAlarmVibratePage';
-import WakeUpAlarmDetail from './src/pages/WakeUpAlarmDetail';
-import WakeUpAlarmPage from './src/pages/WakeUpAlarmPage';
-import WakeUpAlarmPageNotAuto from './src/pages/WakeUpAlarmPageNotAuto';
 
 // 화면 높이 가져오기
 const screenHeight = Dimensions.get('window').height;
