+++ Time-Up/src/pages/Alarm/AlarmPage.tsx
@@ -1,7 +1,7 @@
 // src/pages/AlarmPage.tsx
 import React, { useEffect } from 'react';
 import { View } from 'react-native';
-import useAppNavigation from '../hooks/useAppNavigation';
+import useAppNavigation from '../../hooks/useAppNavigation';
 
 export default function AlarmPage() {
   const navigation = useAppNavigation();
